Move define_omero_channels to specific module
